refactor: changing code to use exception handler

+++ app/usecases/categories/delete_category.py
@@ -2,6 +2,7 @@ from app.domain.abstractions.usecases import AbstractUseCase
 from app.domain.abstractions.repositories import (
     AbstractCategoryRepository,
 )
+from app.domain.exceptions.categories import CategoryNotFoundException
 
 
 class DeleteCategoryUseCase(AbstractUseCase):
@@ -16,6 +17,10 @@ class DeleteCategoryUseCase(AbstractUseCase):
         self,
         category_id: int,
     ) -> None:
+        
+        category = self.category_repo.get_category_by_id(category_id=category_id)
+        if not category:
+            raise CategoryNotFoundException(f"Category with id {category_id} not found")
         
         return self.category_repo.delete(
             category_id=category_id,
