+++ app/usecases/transactions/delete_transaction.py
@@ -3,6 +3,8 @@ from app.domain.abstractions.repositories import (
     AbstractTransactionRepository,
     AbstractUserRepository,
 )
+from app.domain.exceptions.transactions import TransactionNotFoundException
+from app.domain.exceptions.users import UserNotFoundException
 
 
 class DeleteTransactionUseCase(AbstractUseCase):
@@ -23,7 +25,11 @@ class DeleteTransactionUseCase(AbstractUseCase):
         
         user_id = self.user_repo.get_user_id_by_username(username=username)
         if not user_id:
-            raise ValueError("User not found")
+            raise UserNotFoundException(f"User {username} not found")
+        
+        transaction = self.transaction_repo.fetch_one(transaction_id=transaction_id)
+        if not transaction:
+            raise TransactionNotFoundException(f"Transaction with id {transaction_id} not found")
         
         return self.transaction_repo.delete(
             transaction_id=transaction_id,
